Basestation's RF module is now deactivated and a link is established
to a Adafruit LoRa Feather that does the actual wireless transceiving.

+++ firmware/src/mac/util.h
@@ -17,7 +17,7 @@
 /**
  * @Source http://stackoverflow.com/questions/10564491/function-to-calculate-a-crc16-checksum/23726131#23726131
  */
-static inline uint16_t crc16(const uint8_t* data_p, uint8_t length) {
+static inline uint16_t crc16(const uint8_t* data_p, uint16_t length) {
     uint8_t x;
     uint16_t crc = 0xFFFF;
 
@@ -32,7 +32,7 @@ static inline uint16_t crc16(const uint8_t* data_p, uint8_t length) {
     return crc;
 }
 
-static inline void mac_frame_calc_crc(uint8_t* frame, size_t frame_size) {
+static inline void mac_frame_calc_crc(uint8_t* frame, uint16_t frame_size) {
     uint16_t crc = crc16((const uint8_t*)frame, frame_size - 2);
     frame[frame_size-2] = (crc >> 8) & 0xFF;
     frame[frame_size-1] = (crc     ) & 0xFF;
